backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
mcp
openai
python-dotenv
